requirements.txt
adbutils==1.0.10
alabaster==0.7.12
allure-pytest==2.11.1
allure-python-commons==2.11.1
apkutils2==1.0.0
Appium-Python-Client==2.7.1
astroid==2.12.11
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.10.3
black==22.1.0
bleach==5.0.1
blue==0.9.1
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cigam==0.0.3
click==8.1.3
colorama==0.4.5
common==0.1.2
commonmark==0.9.1
configparser==5.3.0
coverage==6.5.0
cryptography==37.0.2
decorator==5.1.1
Deprecated==1.2.13
deprecation==2.1.0
dill==0.3.5.1
distlib==0.3.6
doc8==1.0.0
docutils==0.19
et-xmlfile==1.1.0
execnet==1.9.0
facebook-wda==1.4.6
filelock==3.8.0
flake8==4.0.1
-e git+https://github.com/grantjenks/free-python-games.git@b593e16c48dc0e2250c45dd0ce2893d0937f8b97#egg=freegames
h11==0.13.0
idna==3.3
imagesize==1.4.1
importlib-metadata==5.0.0
iniconfig==1.1.1
isort==5.10.1
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.9.3
lazy-object-proxy==1.7.1
logzero==1.7.0
lxml==4.9.1
MarkupSafe==2.1.1
mccabe==0.6.1
more-itertools==8.14.0
mypy-extensions==0.4.3
openpyxl==3.0.10
outcome==1.1.0
packaging==20.9
pathspec==0.10.1
pbr==5.10.0
Pillow==9.3.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
progress==1.6
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyelftools==0.29
pyflakes==2.4.0
pygame==2.1.2
Pygments==2.13.0
pylint==2.15.4
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
pypiwin32==223
PySocks==1.7.1
pytest==7.1.2
pytest-cov==4.0.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-xdist==3.0.2
pytz==2022.4
PyUserInput==0.1.10
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
readme-renderer==37.2
requests==2.27.1
requests-toolbelt==0.10.0
restructuredtext-lint==1.4.0
retry==0.9.2
rfc3986==2.0.0
rich==12.6.0
rstcheck==3.5.0
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==5.2.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stevedore==4.0.1
tomli==2.0.1
tomlkit==0.11.5
tornado==6.2
tox==3.26.0
trio==0.20.0
trio-websocket==0.9.2
twine==4.0.1
typing_extensions==4.4.0
uiautomator2==2.16.19
undetected-chromedriver==3.1.6
urllib3==1.26.9
virtualenv==20.16.5
webencodings==0.5.1
websockets==10.4
weditor==0.6.4
whichcraft==0.6.1
wrapt==1.14.1
wsproto==1.1.0
xmltodict==0.13.0
zipp==3.9.0
